requirements.txt
Django==3.2.6
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
django-cors-headers==3.7.0
drf-yasg==1.20.0
gunicorn==20.1.0
pandas==1.3.0
reportlab==3.6.12
psycopg2-binary==2.9.3
